# perf(native-filters): improve native filter modal form performance