--- a/superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FilterConfigurePane.tsx
+++ b/superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FilterConfigurePane.tsx
@@ -22,7 +22,7 @@ import FilterTitlePane from './FilterTitlePane';
 import { FilterRemoval } from './types';
 
 interface Props {
-  children: (filterId: string) => React.ReactNode;
+  children: React.ReactNode;
   getFilterTitle: (filterId: string) => string;
   onChange: (activeKey: string) => void;
   onAdd: (type: NativeFilterType) => void;
@@ -62,40 +62,24 @@ const FilterConfigurePane: React.FC<Props> = ({
   currentFilterId,
   filters,
   removedFilters,
-}) => {
-  const active = filters.filter(id => id === currentFilterId)[0];
-  return (
-    <Container>
-      <TitlesContainer>
-        <FilterTitlePane
-          currentFilterId={currentFilterId}
-          filters={filters}
-          removedFilters={removedFilters}
-          erroredFilters={erroredFilters}
-          getFilterTitle={getFilterTitle}
-          onChange={onChange}
-          onAdd={(type: NativeFilterType) => onAdd(type)}
-          onRearrange={onRearrange}
-          onRemove={(id: string) => onRemove(id)}
-          restoreFilter={restoreFilter}
-        />
-      </TitlesContainer>
-      <ContentHolder>
-        {filters.map(id => (
-          <div
-            key={id}
-            style={{
-              height: '100%',
-              overflowY: 'auto',
-              display: id === active ? '' : 'none',
-            }}
-          >
-            {children(id)}
-          </div>
-        ))}
-      </ContentHolder>
-    </Container>
-  );
-};
+}) => (
+  <Container>
+    <TitlesContainer>
+      <FilterTitlePane
+        currentFilterId={currentFilterId}
+        filters={filters}
+        removedFilters={removedFilters}
+        erroredFilters={erroredFilters}
+        getFilterTitle={getFilterTitle}
+        onChange={onChange}
+        onAdd={(type: NativeFilterType) => onAdd(type)}
+        onRearrange={onRearrange}
+        onRemove={(id: string) => onRemove(id)}
+        restoreFilter={restoreFilter}
+      />
+    </TitlesContainer>
+    <ContentHolder>{children}</ContentHolder>
+  </Container>
+);
 
 export default FilterConfigurePane;