--- a/superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigForm/utils.ts
+++ b/superset-frontend/src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigForm/utils.ts
@@ -38,7 +38,7 @@ export const FILTER_SUPPORTED_TYPES = {
   filter_range: [GenericDataType.NUMERIC],
 };
 
-export const useForceUpdate = (isActive: boolean) => {
+export const useForceUpdate = (isActive = true) => {
   const [, updateState] = React.useState({});
   return React.useCallback(() => {
     if (isActive) {