<!---
Please write the PR title following the conventions at https://www.conventionalcommits.org/en/v1.0.0/
Example:
fix(dashboard): load charts correctly
-->

### SUMMARY
<!--- Describe the change below, including rationale and design decisions -->
Improve native filter modal performance by adding memo or reduce rendering.
### BEFORE/AFTER SCREENSHOTS OR ANIMATED GIF
<!--- Skip this if not applicable -->
Test case : 30+ native filters, 20+ charts

| - | before | after |
| - | - | - |
| re-rendering(counts) | 538 |  4 |
| opening time | 15s | 300 ms |
|  flame graph | <img width="1253" alt="image" src="https://user-images.githubusercontent.com/11830681/196002050-aa5c8dfb-097d-4aef-a194-8d4dd1299ae0.png"> |  <img width="490" alt="image" src="https://user-images.githubusercontent.com/11830681/196001864-c2492d8c-72cc-4538-8beb-7199107fedb1.png"> |

### before
https://user-images.githubusercontent.com/11830681/196002347-dad0dfe9-2aef-4107-a5a0-6118d56dbfb6.mov

### after


https://user-images.githubusercontent.com/11830681/196002478-ab260806-de9f-4ca9-98af-5d48a5da8a9a.mov


### TESTING INSTRUCTIONS
<!--- Required! What steps can be taken to manually verify the changes? -->

### ADDITIONAL INFORMATION
<!--- Check any relevant boxes with "x" -->
<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->
- [x] Has associated issue: https://github.com/apache/superset/issues/21147
- [ ] Required feature flags:
- [ ] Changes UI
- [ ] Includes DB Migration (follow approval process in [SIP-59](https://github.com/apache/superset/issues/13351))
  - [ ] Migration is atomic, supports rollback & is backwards-compatible
  - [ ] Confirm DB migration upgrade and downgrade tested
  - [ ] Runtime estimates and downtime expectations provided
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API